lambda/requirements.txt
ask-sdk-core>=1.11.0
airtable-python-wrapper>=0.15.2